Refactor: move isBinOp to more general util class
